+++ src/main/kotlin/no/nav/familie/pdf/pdf/pdfElementer/Seksjoner.kt
@@ -2,8 +2,10 @@ package no.nav.familie.pdf.pdf.pdfElementer
 
 import com.itextpdf.kernel.colors.DeviceRgb
 import com.itextpdf.kernel.pdf.canvas.draw.SolidLine
+import com.itextpdf.layout.Document
 import com.itextpdf.layout.element.Div
 import com.itextpdf.layout.element.LineSeparator
+import no.nav.familie.pdf.pdf.domain.FeltMap
 import no.nav.familie.pdf.pdf.domain.VerdilisteElement
 import no.nav.familie.pdf.pdf.visningsvarianter.håndterVisningsvariant
 
@@ -50,5 +52,13 @@ fun håndterRekursivVerdiliste(
                 }
             }
         }
+    }
+}
+
+fun Document.leggTilSeksjoner(
+    feltMap: FeltMap,
+) {
+    feltMap.verdiliste.forEach {
+        add(lagSeksjon(it))
     }
 }
